402 and 422

+++ src/main/java/com/LSG/DESAFIO/CRUD/de/clientes/services/ClientService.java
@@ -3,7 +3,9 @@ package com.LSG.DESAFIO.CRUD.de.clientes.services;
 import com.LSG.DESAFIO.CRUD.de.clientes.DTO.ClientDTO;
 import com.LSG.DESAFIO.CRUD.de.clientes.entities.Client;
 import com.LSG.DESAFIO.CRUD.de.clientes.repositories.ClientRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -21,6 +23,7 @@ public class ClientService {
                 ()->new ResourceNotFoundException("Recurso não encontrado"));
         return new ClientDTO(client);
     }
+
     @Transactional(readOnly = true)
     public Page<ClientDTO> findAll(Pageable pageable){
         Page<Client> clients =  repository.findAll(pageable);
@@ -29,6 +32,7 @@ public class ClientService {
 
     @Transactional
     public ClientDTO insert(ClientDTO dto){
+
         Client client = new Client();
 
         client.setName(dto.getName());
@@ -44,20 +48,32 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(ClientDTO dto , Long id){
-        Client client = repository.getReferenceById(id);
+        try {
+            Client client = repository.getReferenceById(id);
 
-        client.setName(dto.getName());
-        client.setChildren(dto.getChildren());
-        client.setCpf(dto.getCpf());
-        client.setIncome(dto.getIncome());
-        client.setBirthDate(dto.getBirthDate());
+            client.setName(dto.getName());
+            client.setChildren(dto.getChildren());
+            client.setCpf(dto.getCpf());
+            client.setIncome(dto.getIncome());
+            client.setBirthDate(dto.getBirthDate());
 
-        repository.save(client);
-        return new ClientDTO(client);
+            repository.save(client);
+            return new ClientDTO(client);
+        }catch (EntityNotFoundException e ){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
     @Transactional
     public void delete(Long id){
-        repository.deleteById(id);
+        if (!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try{
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e ){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
+
     }
 }
